Refactor agent tools and enhance command execution capabilities.
Replace the deprecated Tavily tool with a new structure, integrating glob and grep search tools into the agent's toolset. Update the bash tool to improve command execution with enhanced safety checks and output management. Adjust imports and streamline the tool registry for better organization and usability.

Made-with: Cursor
