+++ src/agent/tools/bash-tool/prompt.ts
